+++ webcrawler/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class TextcrawlingConfig(AppConfig):
+class webcrawlerConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'webcrawler'
