+++ webcrawler/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
-from .views import crawl_website
+from .views import summary
 
 urlpatterns = [
-    path('crawl/', crawl_website, name='crawl_website'),
+    path('summary/', summary, name='summary'),
 ]
